// client/app/models/transaction.ts
export class Transaction {
    transactionId: string;
    stakeholder: any;
    transactionType: string;
    transactionNo: string;
    transactionDescription: string;
    transactionTime: Date;
    isRecorded: boolean;
    createdBy: any;
    updatedBy: any;
    createdAt: Date;
    updatedAt: Date;
}